Created a separate file for tarring logic
